+++ lib/updat_controller.dart
@@ -1,0 +1,1 @@
+export 'package:updat/updat.dart' show UpdatController;